refactor: Consolidate LdAiConfigTypes into a single file
Keep all nested config payload types in one static class instead of
splitting them across partial class files.
